aider: Generate a 10MB file in memory and upload it using the drag and drop function on the page.
# Aider chat conversation:

USER: there are new instructions in the login file

ASSISTANT: I see that there are new instructions in the `login.cy.ts` file. However, I need more specific information about what changes you want to make. Could you please provide more details or instructions?

USER: You task is to generate a 10 mb file. If possible, generate this file in memory. Then upload the file on the page. The page has a "drag and drop" option, try to use this.

ASSISTANT: To make this change we need to modify `cypress/e2e/ohsmart/login.cy.ts` to:

1. Generate a 10MB file in memory.
2. Upload the file using the drag and drop function on the page.

Here is the *SEARCH/REPLACE* block:

cypress/e2e/ohsmart/login.cy.ts
```typescript
<<<<<<< SEARCH
    // generate a 10 mb zip file in memory
    // upload the file using the drag and drop upload function on the page
=======
    // generate a 10 mb file in memory
    const blob = new Blob([new ArrayBuffer(1024 * 1024 * 10)], { type: 'application/zip' });
    const file = new File([blob], 'test.zip', { type: 'application/zip' });

    // upload the file using the drag and drop upload function on the page
    cy.get('input[type=file]').attachFile(file);
>>>>>>> REPLACE
```

This change will generate a 10MB file in memory and upload it using the drag and drop function on the page. Please note that this code assumes that there is an `input[type=file]` element on the page for the file upload. If the actual element is different, you will need to adjust the selector accordingly. Also, this code uses the `attachFile` command from `cypress-file-upload`, which is a third-party plugin for Cypress that provides file upload capabilities. If you haven't installed this plugin yet, you can do so by running `npm install --save-dev cypress-file-upload` and then adding `import 'cypress-file-upload';` to your `commands.js` file.