aider: Attach the zip file using the cypress-file-upload plugin.
# Aider chat conversation:

USER: using cypress-file-upload plugin attach the zip file

ASSISTANT: To make this change we need to modify `cypress/e2e/ohsmart/login.cy.ts` to:

1. Import the `cypress-file-upload` plugin.
2. Use the `attachFile` function from the `cypress-file-upload` plugin to attach the generated file.

Here is the *SEARCH/REPLACE* block:

cypress/e2e/ohsmart/login.cy.ts
```typescript
<<<<<<< SEARCH
    // using cypress-file-upload plugin attach the zip file
=======
    // using cypress-file-upload plugin attach the zip file
    cy.get('input[type=file]').attachFile({
      fileContent: blob.stream(),
      fileName: 'test.zip',
      mimeType: 'application/zip'
    });
>>>>>>> REPLACE
```

This change will use the `attachFile` function from the `cypress-file-upload` plugin to attach the generated file. Please note that this code assumes that there is an `input[type=file]` element on the page for the file upload. If the actual element is different, you will need to adjust the selector accordingly.